Implement specific 'Relationships' model for concept relationships
- Add new Relationship type that holds concept relationships
- Copy code for reading and writing relationships from/to interface{} from the current ontology implementation
- Implement mechanism for serialising and deserialising relationships to json correctly. Preserving the flat json messaging format.

The logic behind this refactoring is that we want to have concept property that stored the relationships for each concept of type 'Relationship'. Basically copy the functionality from current ontology lib implementation.
But at the same time we want to keep the messaging format as they were before the ontology extraction. As in all properties and relationships were in a flat json structure.

To accomplish these goals we leverage custom json marshal/unmarshal and substructures.