Remove unnecessary code from the sqs package
The `sqs` package won't be used to SendEvents anymore. Instead AGGY will use
SNS to publish events. There is no point in keeping logic and models
which won't be used.

Also previously it was possible to create a new sqs client with empty
queueURL because in US there isn't a concepts events queue.