Migrate service to cm-graph-ontology V2
The main impact is that now when aggregating concepts the canonical concept contains aggregated relationships of all source concepts.
Also the name of NewAggregatedConcept and NewConcept structs are now CanonicalConcept and SourceConcept respectively. The other underlining logic is the same.

Test were modified to use TestConcept type as to not break them.